Implement the scanner_argument_parser
This class serves as an indirection layer over the argparser library and
encapsulates the CLI command definition and the translation of the raw
user input into the known set of arguments.

The ideia in the near future is for this class to be able to pre-process
all the CLI input into some sort of request object for the scanner, in
order to decouple this task from the main scanner function which
currently is one of the reasons that prevents the function to be used in
any other context besides a CLI run.