Added the beginnings of the changes needed for the setup.py to work.
Changed around the directory structure to have the files a part of the VHostScan
package so it could be used for pkg_resources. Also changed the DEFAULT_WORDLIST_FILE to
use pkg_resources instead of __file__. Also changed some of the imports so they are using relative
imports instead of the full path.

This appears to work with atleast the -t flag, however, more testing should be done and the tests that
already exist should be able to run when running `python setup.py test`